// src/app/site/interfaces/site-cv.interface.ts
// Generated by https://quicktype.io

export interface SiteCv {
  personal:    Personal;
  experiencia: Experiencia;
  formacion:   Formacion;
  cursos: Formacion;
  capacidades: Capacidades;
}

// ****** Personal ******
  export interface Personal {
    titulo: string;
    urlFoto?:      string;
    labelNombre?:  string;
    nombre:       string;
    labelTelefono:string;
    telefono:     string;
    labelMail:    string;
    mail:         string;
    labelNacionalidad: string;
    nacionalidad: string;
    extra: PersonalDatoExtra[];
  }

  export interface PersonalDatoExtra{
    label: string;
    value: string;
  }

// ****** Experiencia *****
  export interface Experiencia {
    titulo: string;
    datos:  ExperienciaDato[];
  }
  export interface ExperienciaDato {
    fechaIni?:    string;
    fechaFin:    string;
    puesto:      string;
    empresa:     string;
    proyectos?:  Proyecto[];
    descripcion?: string[];
    tecnologias?: string;
  }
  export interface Proyecto{
    nombre: string;
    fechaIni?: string;
    fechaFin?: string;
    descripcion: string[];
    tecnologias?: string;
  }

// ****** Formacion *****
  export interface Formacion {
    titulo: string;
    datos:  FormacionDato[];
  }
  export interface FormacionDato {
    fechaIni?:    string;
    fechaFin:    string;
    horas?:       number;
    nombreCentro?:string;
    titulo:      string;
    descripcion: string;
  }

// ****** Capacidades ******
  export interface Capacidades {
    titulo:       string;
    labelNativo:       string;
    nativo: string;
    labelOtrosIdiomas?: string;
    labelLectura?: string;
    labelEscritura?: string;
    labelOral?: string;
    idiomas:      Idioma[];
    labelComunicacion: string;
    comunicacion: string;
    labelOrganizacion: string;
    organizacion: string;
    labelTecnicas?: string;
    tecnicas?:     string;
    labelCarnet?: string;
    carnet?:       string;
  }
  export interface Idioma {
    titulo:    string;
    lectura:   LangLevel;
    escritura: LangLevel;
    oral:      LangLevel;
  }

  export enum LangLevel {
    bajo = 'Bajo',
    medio = 'Medio',
    alto = 'Alto',
    muy_alto = 'Muy Alto'
  }